### 2022-10-08 #4828 @Kevin-Mattheus-Moerman
@matteocao the above lines of code counts seem very low :point_up: can you clarify what the submission to be reviewed entails? Is it just the content over at  https://github.com/giotto-ai/giotto-deep or is there other content/repos too?
