### 2022-10-08 #4828 @matteocao
Dear @Kevin-Mattheus-Moerman ,

Thank you for your comments.

If I understand your question correctly, I would say that the `editorialbot` is not counting properly: it is excluding all the files with `.py` extension as well as `.ipynb` which form the core of the library (thousands of lines). Furthermore, there are k8 yaml files not considered either.
The license is the standard AGPLv3, which is also not discovered by the `editorialbot` apparently.

Finally, the **library code** is in the `main` branch of https://github.com/giotto-ai/giotto-deep , while the paper is in the `paper` branch.
Please let me know if this clarifies the submission content.
