### 2022-10-09 #4828 @arfon
Here's the output of `cloc` on the `main` branch:

```
 cloc giotto-deep/
     198 text files.
     196 unique files.
      21 files ignored.

Language                     files          blank        comment           code
Python                         119           1991           3403           7959
Jupyter Notebook                18              0           5264           1894
YAML                            15             18             13            491
Markdown                         5             84              0            313
reStructuredText                19            161            101            250
JSON                             3              0              0             61
Dockerfile                       3             17             10             36
DOS Batch                        1              8              1             26
SVG                              2              0              0             17
make                             1              4              7              9
HTML                             1              0              0              1
SUM:                           187           2283           8799          11057
```
