### 2022-10-11 #4828 @osorensen
👋 @ismailguzel @leotrs @EduPH @Annielytix, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
