### 2022-10-12 #4846 @editorialbot
[REVIEW]: giotto-deep: A Python Package for Topological Deep Learning
**Submitting author:** @matteocao (<a href="http://orcid.org/0000-0001-9416-9090">Matteo Caorsi</a>)
**Repository:** https://github.com/giotto-ai/giotto-deep
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.3
