### 2022-10-13 #4846 @leotrs
Hello @matteocao! The paper has three authors and I see that all three are contributors to the repository. However, there are two contributors that i) are not authors, and ii) have contributed more than one contributor who is an author (at least by GitHub's measures [here](https://github.com/giotto-ai/giotto-deep/graphs/contributors))

Could you please clarify how authorship was determined for your JOSS submission?
