### 2022-10-13 #4846 @matteocao
Dear @leotrs ,
Thank you for your question.
Of the two contributors (that are not authors) you mention, one has joined the efforts very recently and hence has not (yet) contributed to any design/core functionality of the library. The other one, on the other hand, after the initial contributions (1y ago) has completely gone silent and still today we are not able to reach him. This is why we did not include him.
Let me know if this answer is satisfactory enough and/or if you have any suggestion on how to best proceed.
