### 2022-10-13 #4846 @leotrs
Thanks for the clarification. I think the current author list makes sense. Do consider to add the names of other/past contributors in the acknowledgements section of the paper.
