### 2022-10-15 #4846 @leotrs
@matteocao I see no mention of related/similar software in the current paper. Are there absolutely no other options for using topological techniques in deep learning?
