### 2022-10-15 #4846 @matteocao
Dear @leotrs ,
Thank you for your message.

> @matteocao I can see your CI runs pytest on the root directory but wasn't able to find a directory containing a test suite. Could you please point me to it? Thanks!

The **unit tests** are written per module, meaning that there is a folder called `tests` in each module folder. Here some examples:
 - https://github.com/giotto-ai/giotto-deep/tree/master/gdeep/data/datasets/tests
 - https://github.com/giotto-ai/giotto-deep/tree/master/gdeep/data/persistence_diagrams/tests
 - https://github.com/giotto-ai/giotto-deep/tree/master/gdeep/data/preprocessors/tests
 - https://github.com/giotto-ai/giotto-deep/tree/master/gdeep/models/tests
 - ...

The **integration tests** (or maybe even callable e2e tests) are the notebookes themselves: all of those in [here](https://github.com/giotto-ai/giotto-deep/tree/master/examples)

We find this way of organising tests reasonable (as opposed to putting all tests in one single folder). Please let us know if this feels reasonable for you as well.

> @matteocao I see no mention of related/similar software in the current paper. Are there absolutely no other options for using topological techniques in deep learning?

There are indeed no package that try to do what we do. I would like however to mention the only possibly similar one, though much more limited in scope: https://github.com/MathieuCarriere/perslay. This packages implements one single technique that is topology-flavoured for feed-forward deep networks. **Important note:** even though we understood the paper and later were able to reproduce its content, we have not been able to use ay part of this "PersLay" package due to the total lack of documentation and structure. Please feel free to advise on what's the best course of action. Thank you!
