### 2022-10-17 #4846 @EduPH
Congratulations @matteocao, it is a nice library for the TDA community. I think the paper can be accepted. Just a remark, it would be nice to mention in the paper that previously, people have used TDA libraries such as Gudhi together with deep learning libraries such as scikit-learn, TensorFlow or PyTorch.
