### 2022-10-24 #4846 @ismailguzel
I tested the package installation on both Windows and Ubuntu. Good for both platforms, actually. I'm grateful for @matteocao
 . I also didn't see anything different from what other reviewers had already noted. Finally, my review is completed, and I think the document can be accepted just as it is.  It will be a great package for TDA folks.
