### 2022-10-24 #4846 @matteocao
Dear @osorensen ,

Thank you for your message. Here the data you asked for:
 - The tagged release version, as there have not been requests to modify the code nor upgrades for the last 20 days, shall be the v0.0.3: https://github.com/giotto-ai/giotto-deep/releases/tag/v0.0.3
 - Zenodo DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7243721.svg)](https://doi.org/10.5281/zenodo.7243721)
 - I doubled checked the data and all is correct.
