### 2022-10-24 #4846 @osorensen
Thanks @matteocao

A few comments on the manuscript

- On lines 28 and 34 you start a sentence with the word `giotto-deep`, but on line 39 you write "Giotto-deep" with capital first letter. Please pick one and be consistent.
- On line 34 I suggest added a "the" in front of `giotto-deep`, so it becomes "The `giotto-deep` architecture ..."
- On lines 66-67 and 73, please add DOI to the arXiv references, cf. https://arxiv.org/abs/2206.15195 and https://arxiv.org/abs/1609.08227

Please let me know when this is done, and I can move forward with accepting the publication.
