+++ paper.bib
@@ -18,6 +18,7 @@
   title={Topological data analysis},
   author={Wasserman, Larry},
   journal={arXiv preprint arXiv:1609.08227},
+  doi={10.48550/arXiv.1609.08227},
   year={2016}
 }
 
@@ -79,5 +80,6 @@
   title={The Topological BERT: Transforming Attention into Topology for Natural Language Processing},
   author={Perez, Ilan and Reinauer, Raphael},
   journal={arXiv preprint arXiv:2206.15195},
+  doi={10.48550/arXiv.2206.15195},
   year={2022}
 }
