### 2022-11-07 #4846 @matteocao
**Thank you All!**

I really love this approach to peer reviews: fully transparent, useful, serious and friendly at the same time. I think JOSS is really bringing a breath of fresh air into the world of publications and imho many publishers should follow this virtuous example!

Congrats for the great journal!

Cheers! 🥂
